refactor: Moving CSS into its own directory
