+++ src/components/Hue.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react";
 
 import Interactive, { Interaction } from "./Interactive";
 
-import styles from "../styles.css";
+import styles from "../css/styles.css";
 import { hsvToHslString } from "../utils/convert";
 import formatClassName from "../utils/formatClassName";
